organize source files
