+++ checkout.html
@@ -487,6 +487,6 @@
     <!-- Template Javascript -->
     <script src="js/main.js"></script>
     <script type="module" src="js/checkout.js"></script>
-    <script type="module" src="header.js"></script>
+    <script type="module" src="js/header.js"></script>
   </body>
 </html>
