+++ index.html
@@ -552,7 +552,7 @@
 
     <!-- Template Javascript -->
     <script src="js/main.js"></script>
-    <script type="module" src="header.js"></script>
-    <script type="module" src="home.js"></script>
+    <script type="module" src="js/header.js"></script>
+    <script type="module" src="js/home.js"></script>
   </body>
 </html>
