+++ login.html
@@ -178,7 +178,7 @@
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.bundle.min.js"></script>
     <script src="lib/easing/easing.min.js"></script>
     <script src="lib/owlcarousel/owl.carousel.min.js"></script>
-    <script type="module" src="header.js"></script>
+    <script type="module" src="./js/header.js"></script>
   </body>
-  <script type="module" src="./LoginScript.js"></script>
+  <script type="module" src="./js/LoginScript.js"></script>
 </html>
