+++ shop.html
@@ -626,7 +626,7 @@
 
     <!-- Template Javascript -->
     <script src="js/main.js"></script>
-    <script type="module" src="header.js"></script>
-    <script type="module" src="shop.js"></script>
+    <script type="module" src="./js/header.js"></script>
+    <script type="module" src="./js/shop.js"></script>
   </body>
 </html>
