+++ signup.html
@@ -199,15 +199,15 @@
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.bundle.min.js"></script>
     <script src="lib/easing/easing.min.js"></script>
     <script src="lib/owlcarousel/owl.carousel.min.js"></script>
-    <script type="module" src="header.js"></script>
+    <script type="module" src="js/header.js"></script>
     <p style="width: 500px; margin: auto">
-      .[First Name, Second Name] should be characters only.<br>
-      .[Email] should be valid.<br>
-      .[Password] should be (8) characters long.<br>
+      .[First Name, Second Name] should be characters only.<br />
+      .[Email] should be valid.<br />
+      .[Password] should be (8) characters long.<br />
       and contains at least (3) small characters, (1) capital character,
-      <br> (1) number and (1) special character.<br><br><br>
-      
+      <br />
+      (1) number and (1) special character.<br /><br /><br />
     </p>
   </body>
-  <script type="module" src="./SignUpScript.js"></script>
+  <script type="module" src="./js/SignUpScript.js"></script>
 </html>
